Move Serialization to utility, to remove completely unnecessary folder.
