+++ Phoenix/Client/Include/Client/Network.hpp
@@ -33,7 +33,7 @@
 #include <Common/Input.hpp>
 #include <Common/Network/Host.hpp>
 #include <Common/Position.hpp>
-#include <Common/Util/BlockingQueue.hpp>
+#include <Common/Utility/BlockingQueue.hpp>
 #include <Common/Voxels/Chunk.hpp>
 
 #include <thread>
