+++ Phoenix/Common/Include/Common/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_subdirectory(Math)
 add_subdirectory(Voxels)
 add_subdirectory(CMS)
-add_subdirectory(Serialization)
+add_subdirectory(Utility)
 add_subdirectory(Network)
 
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
@@ -9,7 +9,7 @@ set(Headers
 	${mathHeaders}
 	${voxelHeaders}
 	${cmsHeaders}
-	${serializationHeaders}
+	${utilityHeaders}
 	${networkHeaders}
 
 	${currentDir}/CoreIntrinsics.hpp
