+++ Phoenix/Common/Include/Common/Utility/CMakeLists.txt
@@ -1,9 +1,11 @@
+add_subdirectory(Internal)
+
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
-set(serializationHeaders
-        ${currentDir}/SharedTypes.hpp
-        ${currentDir}/Endian.hpp
+set(utilityHeaders
+	${currentDir}/BlockingQueue.hpp
+
         ${currentDir}/Serializer.hpp
         ${currentDir}/Serializer.inl
 
         PARENT_SCOPE
-        )
+)
