+++ Phoenix/Common/Include/Common/Utility/Internal/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(currentDir ${CMAKE_CURRENT_LIST_DIR})
+set(utilHeaders
+	${currentDir}/SharedTypes.hpp
+	${currentDir}/Endian.hpp
+	${currentDir}/Front.hpp
+
+        PARENT_SCOPE
+)
